requirements.txt
python-dotenv
requests
fastapi[all]
pipecat-ai[daily,openai,deepgram,elevenlabs,silero] @ git+ssh://git@github.com/photogpt/dubit-pipecat.git@main